refactor: drop briefing-lag transitional re-exports
The transitional re-exports of detectDefaultBranches, defaultRunGit, RunGit,
DetectedBranches, and RunGitResult were added during the git-runner.ts
extraction (PR #366) so existing importers did not have to churn. Update
the three remaining importers (src/briefing-lag.test.ts,
src/staging-ff.test.ts, src/mag.test.ts) to pull these symbols from
./git-runner.ts directly, then remove the re-export block from
briefing-lag.ts. formatUpstreamLagSection / parseLeftRightCount continue to
live in briefing-lag.ts; src/mag.ts's import of formatUpstreamLagSection is
unaffected.
